+++ deltafi-ui/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+-  Ability to repopulate metadata on DeltaFi Upload page. (#125)
+
 ## [0.19.0]
 
 ### Added